Decides exclusions based on plaintext paths.
Since exclusions will support wildcards, it's not possible to store
the list of exclusions as a list of encrypted paths (calculated during
initialization). Instead, the plain paths are stored, and the path to
be checked is decrypted at the time of isExcluded check and compared
with the stored plaintext paths.