Exclude DirIV files if parent directory is excluded.
Assuming a directory SOMEDIR (possibly deep in the filesystem
hierarchy) is excluded, now SOMEDIR/gocryptfs.diriv is considered
excluded.

If it were not excluded, it might be possible to infer the existence
of SOMEDIR by checking for SOMEDIR/gocryptfs.diriv.

In practice I've not been able to actually exploit this, as fuse or
the kernel tries to open each directory in a path before actually
trying to open (or stat) a file, but it's safer if gocryptfs makes a
further check.