Logging corrections...
Logging, since we're on Linux/OSX/*BSD, should just be
a matter of switching the writer for the logging in tlog
and the default logger to the syslog one.  We need to,
however, do a .Dial call because .New appears to not
do what one intends- it doesn't establish a connection
to the syslog daemon unless you already have one from
.Dial

So, with that, we cleaned out the piped shell call to
logger, which adds to the overhead, isn't needed and
open up the UNIX domain socket hook to syslog with
log/syslog instead.  Seems to work for the cases
I needed for things.  Future iterations can accept
an option for -tcp/-udp and an IP address for
a network logging option if needed now.