changed badnames mechanic to regexp (test)
badnames parameter expects now pattern for regexp package. Pattern is applied to cipher file/directory names if direct decryption fails. If pattern generates a match, the match is removed from cipher name and the decryption is applied again. If this produces a valid name, this name is display and the pattern match and a bad name indicator concatenated.