fusefronted: fake fstat for deleted fifos
git test suite t9300-fast-import.sh test 245 does the equivalent of this:

	mkfifo fifo
	exec 8<>fifo
	rm fifo
	cat /dev/null >&8

This used to fail with

	cat: standard output: No such file or directory

because cat tries to fstat stdout.

The open() on the fifo does not reach the filesystem, so we don't
have an fd open for the delted file, hence no way to access it.

Fake fstat success as good as we can to make cat happy.

Fixes https://github.com/rfjakob/gocryptfs/issues/929